apps/mysql_server/requirements.txt
confluent-kafka
pydantic
mysql-connector-python
pydantic-settings
